Manually poke at dbus on linux
It seems that inhibiting the screensaver on Linux only persists as long
as the calling process is alive. Because `xdg-screensaver` is a shell
script that uses `dbus-send` to send a message, the calling process is
actually `dbus-send`, so when it terminates, the inhibition expires.

At least that's what I think, after lots of trawling the googles. I'm
making essentially the exact same dbus calls, but it works when I do it
from this library, and not when `xdg-screensaver` sends them.

I don't know that this is as distro-agnostic as I'd like it to be. I
didn't want to replicate the unholy mess that is `xdg-screensaver`.