Abuse the extension system for platform dependencies.
A standard `gemspec` won't let us specify dependencies by platform. In
our case, though, we only want to install `ruby-dbus` on Linux. This
pretends to compile a native extension, and just installs the gem if
on the right platform.

Two problems:

* This is non-intuitive and not clear what dependencies exist.
* This will fail (for no real reason) if rake isn't installed.

Is this a good idea? Is there a better way? Should we just say fuggit
and install `ruby-dbus` on all platforms?