Perform an async DBus call
Under some circumstances inhibiting the screen saver will block (e.g. a
system with the org.freedesktop.ScreenSaver service provided by Cinnamon
but the current window-manager is not Cinnamon).

Do the call asynchroneously so that fidget is not blocking, and Ctrl+C
will be catched propertly and configuration reset as expected in all
cases.